feat(debug): add debugRun and modify main according to release mode
